+++ NEWS.md
@@ -8,6 +8,8 @@
 
 - a new function `all_labels()` to get all chunk labels in a document; see one application at https://github.com/yihui/knitr-examples/blob/master/073-code-appendix.Rnw
 
+- chunk hooks will be run (if exist) even if the code chunk is empty
+
 ## BUG FIXES
 
 - fixed #432: no longer uses `\\\\` in LaTeX output; only a single line break is converted to `\\` (thanks, Kevin Wright)